Update indexing.md
My goal here is to document that .ppignore does not ignore already indexed directories or files.  If you wish to ignore such directories or files the only method is to create the .ppignore file, reset the database and reindex.
The only caveat with this documentation is that it would be best to also disclose any risks associated with running photoprism reset.  I don't know what risks are associated, so I can't add further there.